requirements.txt
django
django-allauth
python-dotenv
django-stubs
django-stubs-ext
django-countries
mypy-extensions
typing_extensions
requests
jwt
psycopg[binary]
python-dotenv
